test/wpt/jsdom/issues/3818.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>3818</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/jsdom/issues/3818 -->

<div id=":r3:"></div>

<script>
"use strict";
const div = document.querySelector('div');

test(() => {
  assert_true(div.matches(':scope'));
}, ':scope pseudo class');
test(() => {
  assert_true(div.matches('#\\:r3\\:'));
}, 'escaped ID');
</script>
